Reverting fix to https://github.com/gorhill/uBlock/issues/3639
Related feedback:
- https://github.com/uBlockOrigin/uAssets/commit/a54cb2e38f7b2af972f2ddee8fd86350d6f2e07c#commitcomment-34387041

Regression from converting uBO resources into
immutable resources. This affects only Chromium-based
browsers.

This is a quick fix for the dev build, to at least unbreak
the reported case.

I need to research whether the root issue (conflict with HTTPS-E)
is still occurring, and if so the fix is not trivial: I will need
to add code to uBO to fetch and convert binary data into `data:`
URIs.