Sixth pass to re-design the dashboard
Related commits:
- https://github.com/gorhill/uBlock/commit/9e775e8864560fde16927bbc22ea4a2e7d572216
- https://github.com/gorhill/uBlock/commit/98bcf0aafb086eba7307b60c06088cdd68f9db7d
- https://github.com/gorhill/uBlock/commit/da29004c40a95794cb26913a8800bcfec599171b
- https://github.com/gorhill/uBlock/commit/e91a62349521246c59c29d9404767f7b6338b807
- https://github.com/gorhill/uBlock/commit/88b63c2ea57eb5d6e9b10bcdb737f78c890ca33b

This commit focuses on implementing the use of
CSS `var(...)` throughout so as to make it easy
to create themes -- a requirement for uBO to
support a dark theme.

There is still work to do regarding converting
uBO's CSS to completely support `var(...)` but
being able to start using theming will help
complete and fine tune CSS `var(...)` support.

Related issue:
- https://github.com/uBlockOrigin/uBlock-issues/issues/401