Add widget to control selector depth to element picker
Further iterating on the work done in following commit:
- https://github.com/gorhill/uBlock/commit/39bf9f466094ba11393a26644cb62a71d67c9d82

This commit adds a new widget to the element picker to
control the depth of a cosmetic filter selector. The
new widget is essentially just another way of selecting
the depth, which is still controllable through picking
one of the cosmetic filters in the list of candidates.