Enable origin-hit coalescing optimisation for modifier filters
Related commit:
- https://github.com/gorhill/uBlock/commit/8466dd44c41d21029bd6196ffa71d318aee77bb1

The optimization in the commit above was meant to
improve the performance of lookup operations of
modifier filters, but I forgot to enable the
optimisation for that class of filters.

This means this commit brings another significant
performance gain on top of the previous commit, as
shown by the built-in benchmark.

Additionally a few minor code rearrangements.